Create Unit test for Employer Api

+++ src/main/java/vn/unigap/api/employer/repository/JobProvinceRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import vn.unigap.api.employer.entity.jpa.JobProvince;
 
-public interface JobProvinceRepository extends JpaRepository<JobProvince, Integer> {
+public interface JobProvinceRepository extends JpaRepository<JobProvince, Long> {
 }
